+++ BTCPayServer/Components/NotificationsDropdown/Default.cshtml
@@ -3,6 +3,7 @@
 @inject ISettingsRepository SettingsRepository
 @using BTCPayServer.HostedServices
 @using BTCPayServer.Views.Notifications
+@using BTCPayServer.Abstractions.Extensions
 @using BTCPayServer.Abstractions.Contracts
 @using BTCPayServer.Client
 @using BTCPayServer.Services
