+++ BTCPayServer/Views/Apps/_ViewImports.cshtml
@@ -1,2 +1,3 @@
-﻿@using BTCPayServer.Views.Apps
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views.Apps
 @using BTCPayServer.Models.AppViewModels
