+++ BTCPayServer/Views/CoinSwitch/UpdateCoinSwitchSettings.cshtml
@@ -1,5 +1,5 @@
-﻿@using Microsoft.AspNetCore.Mvc.Rendering
-@using BTCPayServer.Views.Stores
+﻿@using BTCPayServer.Views.Stores
+@using BTCPayServer.Abstractions.Extensions
 @model UpdateCoinSwitchSettingsViewModel
 @{
     Layout = "../Shared/_NavLayout.cshtml";
