+++ BTCPayServer/Views/EthereumLikeStore/GetStoreEthereumLikePaymentMethod.cshtml
@@ -1,4 +1,5 @@
 @using BTCPayServer.Views.Stores
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Services.Altcoins.Ethereum.UI.EditEthereumPaymentMethodViewModel
 
 @{
