+++ BTCPayServer/Views/EthereumLikeStore/GetStoreEthereumLikePaymentMethods.cshtml
@@ -1,5 +1,6 @@
 
 @using BTCPayServer.Views.Stores
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Services.Altcoins.Ethereum.UI.ViewEthereumStoreOptionsViewModel
 @inject SignInManager<ApplicationUser> SignInManager;
 @inject BTCPayNetworkProvider BTCPayNetworkProvider;
