+++ BTCPayServer/Views/Invoice/_ViewImports.cshtml
@@ -1,2 +1,3 @@
-﻿@using BTCPayServer.Services.Invoices
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Services.Invoices
 @using BTCPayServer.Views.Invoice
