+++ BTCPayServer/Views/Manage/_ViewImports.cshtml
@@ -1,1 +1,2 @@
-﻿@using BTCPayServer.Views.Manage
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views.Manage
