+++ BTCPayServer/Views/Manage/_ViewStart.cshtml
@@ -1,4 +1,5 @@
-﻿@using BTCPayServer.Views
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views
 @using BTCPayServer.Views.Manage
 @{
     Layout = "../Shared/_NavLayout.cshtml";
