+++ BTCPayServer/Views/MoneroLikeStore/GetStoreMoneroLikePaymentMethod.cshtml
@@ -1,5 +1,5 @@
-@using BTCPayServer.Controllers
 @using BTCPayServer.Views.Stores
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Services.Altcoins.Monero.UI.MoneroLikeStoreController.MoneroLikePaymentMethodViewModel
 
 @{
