+++ BTCPayServer/Views/Notifications/_ViewImports.cshtml
@@ -1,1 +1,1 @@
-﻿
+﻿@using BTCPayServer.Abstractions.Extensions
