+++ BTCPayServer/Views/PaymentRequest/_ViewImports.cshtml
@@ -1,1 +1,2 @@
-﻿@using BTCPayServer.Views.PaymentRequest
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views.PaymentRequest
