+++ BTCPayServer/Views/PaymentRequest/_ViewStart.cshtml
@@ -1,3 +1,4 @@
+@using BTCPayServer.Abstractions.Extensions
 @using BTCPayServer.Views
 @using BTCPayServer.Views.PaymentRequest
 @{
