+++ BTCPayServer/Views/Server/_ViewImports.cshtml
@@ -1,2 +1,3 @@
-﻿@using BTCPayServer.Views.Server
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views.Server
 @using BTCPayServer.Models.ServerViewModels
