+++ BTCPayServer/Views/Shared/Ethereum/UpdateChainConfig.cshtml
@@ -1,6 +1,6 @@
-@using BTCPayServer.Services.Altcoins.Ethereum.UI
 @using BTCPayServer.Views.Server
 @using System.Net.Http
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Services.Altcoins.Ethereum.Configuration.EthereumLikeConfiguration
 @inject BTCPayNetworkProvider BTCPayNetworkProvider;
 @inject IHttpClientFactory HttpClientFactory;
