+++ BTCPayServer/Views/UserStores/_ViewImports.cshtml
@@ -1,2 +1,3 @@
-﻿@using BTCPayServer.Views.Stores
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views.Stores
 @using BTCPayServer.Models.StoreViewModels
