+++ BTCPayServer/Views/Wallets/_ViewImports.cshtml
@@ -1,3 +1,4 @@
-﻿@using BTCPayServer.Views.Wallets
+﻿@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Views.Wallets
 @using BTCPayServer.Models.WalletViewModels
 @addTagHelper *, BundlerMinifier.TagHelpers
